coinomi launch date
As stated here https://twitter.com/CoinomiWallet/status/1209074029497344002
the wallet was started in January 2014. Full tweet:
That's much better, thank you for the amendments. Also, we started operating on January 2014, not November. Our account was reset but the app went live during the first days of 2014. Another thing to note is that an app can be open source but the binary uploaded to the stores