show author, show altTitle everywhere
Reviewers should get credit, also to make clear that the analysis is not
done by an anonymous organization. Future experts may disagree with
authors, too.
altTitle is not an after-thought. It's shown now everywhere and
explained in the analysis.
the linux template was removed. It got added accidentally and is not
worth maintaining on master.