Added drawStar and drawStarFunction
i'm pushing the star function - with your refactoring changes and without changing the font of the verdict.
There is some math involved in computing the 'partial stars' but I can't get it right.
The idea is if an app has 2.5 stars, it should draw 2 full stars and 0.5 of a star.
Now the approach to take is to draw a white rectangle spanning half of the star. Thus producing a half star starting from the right.